Updating our branches ahead of documentation push

+++ plugin/qemu-memory.h
@@ -8,7 +8,7 @@
  * Approved for Public Release, Distribution Unlimited
  *
  * Authors:
- *  Adam Critchley <adamc@cromulence.com>
+ *  Adam Critchley <shoggoth@cromulence.com>
  *
  * This work is licensed under the terms of the GNU GPL, version 2 or later.
  * See the COPYING file in the top-level directory.
@@ -25,8 +25,8 @@
 extern "C"{
 #endif
 
-bool qemu_get_virtual_memory(int cpu_id, uint64_t address, uint8_t size, uint8_t **data);
-bool qemu_set_virtual_memory(int cpu_id, uint64_t address, uint8_t size, uint8_t *data);
+bool qemu_get_virtual_memory(int cpu_id, uint64_t address, uint64_t size, uint8_t **data);
+bool qemu_set_virtual_memory(int cpu_id, uint64_t address, uint64_t size, uint8_t *data);
 void qemu_get_physical_memory(uint64_t address, uint64_t size, uint8_t **data);
 void qemu_set_physical_memory(uint64_t address, uint64_t size, uint8_t *data);
